third_party/blink/web_tests/http/tests/security/no-popup-from-sandbox-top.html

<script>
if (window.testRunner) {
    testRunner.dumpAsText();
    testRunner.setPopupBlockingEnabled(false);
}
</script>
<p>To run this test outside of DumpRenderTree, please disable your popup blocker!</p>
<iframe sandbox="allow-scripts"
  src="data:text/html,
       <script>
       var win = window.open('about:blank', '_top');
       console.log(win ? 'FAIL' : 'PASS');
       </script>"
  ></iframe>
